Modify account card on confirmation page
